feat(web-sdk): make reliable Fetch transport the default
