+++ e2e/smoke/export-surface.json
@@ -20,6 +20,7 @@
       "UserActionState": "object",
       "VERSION": "string",
       "allLogLevels": "object",
+      "captureMetas": "function",
       "clampSamplingRate": "function",
       "createInternalLogger": "function",
       "createPromiseBuffer": "function",
@@ -101,6 +102,7 @@
       "UserActionState": "object",
       "VERSION": "string",
       "allLogLevels": "object",
+      "captureMetas": "function",
       "clampSamplingRate": "function",
       "createInternalLogger": "function",
       "createPromiseBuffer": "function",
@@ -182,6 +184,7 @@
       "UserActionState": "object",
       "VERSION": "string",
       "allLogLevels": "object",
+      "captureMetas": "function",
       "clampSamplingRate": "function",
       "createInternalLogger": "function",
       "createPromiseBuffer": "function",
@@ -282,6 +285,7 @@
       "allLogLevels": "object",
       "browserMeta": "function",
       "buildStackFrame": "function",
+      "captureMetas": "function",
       "createInternalLogger": "function",
       "createPromiseBuffer": "function",
       "createSession": "function",
@@ -382,6 +386,7 @@
       "allLogLevels": "object",
       "browserMeta": "function",
       "buildStackFrame": "function",
+      "captureMetas": "function",
       "createInternalLogger": "function",
       "createPromiseBuffer": "function",
       "createSession": "function",
@@ -482,6 +487,7 @@
       "allLogLevels": "object",
       "browserMeta": "function",
       "buildStackFrame": "function",
+      "captureMetas": "function",
       "createInternalLogger": "function",
       "createPromiseBuffer": "function",
       "createSession": "function",
