+++ packages/core/src/api/events/initialize.ts
@@ -1,6 +1,6 @@
 import type { Config } from '../../config';
 import type { InternalLogger } from '../../internalLogger';
-import type { Metas } from '../../metas';
+import { captureMetas, type Metas } from '../../metas';
 import { TransportItemType } from '../../transports';
 import type { TransportItem, Transports } from '../../transports';
 import type { UnpatchedConsole } from '../../unpatchedConsole';
@@ -65,7 +65,7 @@ export function initializeEventsAPI({
       }
 
       const previousPayload = lastPayload;
-      const meta = metas.capture();
+      const meta = captureMetas(metas);
       // Capture can emit a nested event. Preserve dedupe without committing a failed capture.
       if (lastPayload !== previousPayload && !skipDedupe && config.dedupe && deepEqual(testingPayload, lastPayload)) {
         return;
