+++ packages/core/src/api/measurements/initialize.ts
@@ -1,6 +1,6 @@
 import type { Config } from '../../config';
 import type { InternalLogger } from '../../internalLogger';
-import type { Metas } from '../../metas';
+import { captureMetas, type Metas } from '../../metas';
 import { TransportItemType } from '../../transports';
 import type { TransportItem, Transports } from '../../transports';
 import type { UnpatchedConsole } from '../../unpatchedConsole';
@@ -64,7 +64,7 @@ export function initializeMeasurementsAPI({
       }
 
       const previousPayload = lastPayload;
-      const meta = metas.capture();
+      const meta = captureMetas(metas);
       if (lastPayload !== previousPayload && !skipDedupe && config.dedupe && deepEqual(testingPayload, lastPayload)) {
         return;
       }
