+++ packages/web-sdk/src/instrumentations/session/sessionManager/PersistentSessionsManager.ts
@@ -59,7 +59,8 @@ export class PersistentSessionsManager {
     return null;
   }
 
-  updateSession = (): void => this.updateUserSession({ refreshActivity: false });
+  updateSession = ({ refreshActivity = true }: { refreshActivity?: boolean } = {}): void =>
+    this.updateUserSession({ refreshActivity });
 
   recordActivity: (sessionId: string) => void = getUserSessionActivityRecorder({
     fetchUserSession: PersistentSessionsManager.fetchUserSession,
@@ -70,7 +71,7 @@ export class PersistentSessionsManager {
   private init(): void {
     document.addEventListener('visibilitychange', () => {
       if (document.visibilityState === 'visible') {
-        this.updateSession();
+        this.updateSession({ refreshActivity: false });
         const sessionId = faro.api?.getSession()?.id;
         if (sessionId) {
           this.recordActivity(sessionId);
