+++ packages/web-sdk/src/instrumentations/session/sessionManager/VolatileSessionsManager.test.ts
@@ -70,7 +70,7 @@ describe('Volatile Sessions Manager.', () => {
     const nextActivityTimeAfterFiveSeconds = fakeSystemTime + 5000;
     jest.setSystemTime(nextActivityTimeAfterFiveSeconds);
 
-    updateSession();
+    updateSession({ refreshActivity: false });
 
     expect(JSON.parse(mockStorage[STORAGE_KEY]).lastActivity).toBe(fakeSystemTime);
     recordActivity(mockInitialSessionId);
