+++ packages/web-sdk/src/transports/fetch-v2/index.ts
@@ -1,1 +1,2 @@
-export { FetchTransport } from './transport';
+/** @deprecated Use the default FetchTransport export. */
+export { FetchTransport } from '../fetch';
