+++ packages/web-sdk/src/transports/fetch/index.ts
@@ -1,2 +1,2 @@
 export { FetchTransport } from './transport';
-export type { ClockFn, FetchTransportOptions, FetchTransportRequestOptions } from './types';
+export type { ClockFn, FetchTransportOptions, FetchTransportRequestOptions, RandomFn, RetryOptions } from './types';
