+++ packages/web-sdk/src/transports/index.ts
@@ -2,4 +2,4 @@ export { ConsoleTransport } from './console';
 export type { ConsoleTransportOptions } from './console';
 
 export { FetchTransport } from './fetch';
-export type { ClockFn, FetchTransportOptions, FetchTransportRequestOptions } from './fetch';
+export type { ClockFn, FetchTransportOptions, FetchTransportRequestOptions, RandomFn, RetryOptions } from './fetch';
